+++ packages/create-addfox-app/CHANGELOG.md
@@ -1,5 +1,13 @@
 # create-addfox-app
 
+## 0.2.4
+
+### Patch Changes
+
+- fix: recover sourcemap of content/background
+- Updated dependencies
+  - @addfox/pkg-manager@0.2.4
+
 ## 0.2.3
 
 ### Patch Changes