+++ packages/plugins/rsbuild-plugin-extension-monitor/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @addfox/rsbuild-plugin-extension-monitor
 
+## 0.2.4
+
+### Patch Changes
+
+- fix: recover sourcemap of content/background
+- Updated dependencies
+  - @addfox/common@0.2.4
+  - @addfox/core@0.2.4
+
 ## 0.2.3
 
 ### Patch Changes