+++ packages/plugins/rsbuild-plugin-vue/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @addfox/rsbuild-plugin-vue
 
+## 0.2.4
+
+### Patch Changes
+
+- fix: recover sourcemap of content/background
+
 ## 0.2.3
 
 ### Patch Changes